# api/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateNoteRequest:
    properties:
      description:
        example: This is a note description
        minLength: 5
        type: string
      url:
        example: https://example.com
        type: string
      user_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
    required:
    - description
    - url
    type: object
  dto.CreateNoteResponse:
    properties:
      created_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      description:
        example: This is a note description
        type: string
      id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
      updated_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      url:
        example: https://example.com
        type: string
      user_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
    type: object
...
  dto.GetNoteResponse:
    properties:
      created_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      description:
        example: This is a note description
        type: string
      id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
      updated_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      url:
        example: https://example.com
        type: string
      user_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
    type: object
  dto.GetNotesResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/dto.NoteResponse'
        type: array
      paging:
        $ref: '#/definitions/request.Paging'
    type: object
...
  dto.NoteResponse:
    properties:
      created_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      description:
        example: This is a note description
        type: string
      id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
      updated_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      url:
        example: https://example.com
        type: string
      user_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
    type: object
...
  dto.UpdateNoteRequest:
    properties:
      description:
        example: Updated note description
        minLength: 5
        type: string
      note_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
      url:
        example: https://updated-example.com
        type: string
      user_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
    type: object
  dto.UpdateNoteResponse:
    properties:
      created_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      description:
        example: This is a note description
        type: string
      id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
      updated_at:
        example: "2025-06-01T20:50:35.388851+07:00"
        type: string
      url:
        example: https://example.com
        type: string
      user_id:
        example: 0198f10c-98c7-71ab-bc9a-7e148b5ece17
        type: string
    type: object
...
  request.Paging:
    properties:
      current_page:
        type: integer
      from:
        type: integer
      has_next_page:
        type: boolean
      has_previous_page:
        type: boolean
      last_page:
        type: integer
      per_page:
        type: integer
      to:
        type: integer
      total_data:
        type: integer
      total_data_in_current_page:
        type: integer
    type: object
  response.ErrorDetail:
    properties:
      code:
        example: error_code
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: Error message
        type: string
    type: object
  response.Meta:
    properties:
      paging:
        $ref: '#/definitions/request.Paging'
      request_id:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorDetail'
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: sammidev4@gmail.com
    name: Sammi Aldhi Yanto
    url: https://lab-sammi.gitbook.io
  description: RESTful API for notes taking app
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Notes Taking API
  version: "1.0"
paths:
...
  /notes:
    get:
      consumes:
      - application/json
      description: List all notes for the authenticated user with pagination and filtering
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Sort by column (e.g., created_at)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort direction (asc/desc)
        in: query
        name: sort_direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notes listed successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetNotesResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List notes
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Create a new note for the authenticated user
      parameters:
      - description: Note creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Note created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateNoteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - notes
  /notes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a note by its ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Note deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a note
      tags:
      - notes
    get:
      consumes:
      - application/json
      description: Retrieve a specific note by its ID for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Note retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.GetNoteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get a note by ID
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update an existing note for the authenticated user
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Note update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Note updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateNoteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"